[Enhancement #412] Remove FlatTerm and AbstractFullTerm.
