+++ src/main/java/cz/cvut/kbss/termit/persistence/dao/TermDao.java
@@ -35,9 +35,7 @@ import cz.cvut.kbss.termit.event.EvictCacheEvent;
 import cz.cvut.kbss.termit.event.TermReferencesUpdatedEvent;
 import cz.cvut.kbss.termit.event.VocabularyContentModifiedEvent;
 import cz.cvut.kbss.termit.exception.PersistenceException;
-import cz.cvut.kbss.termit.model.AbstractFullTerm;
 import cz.cvut.kbss.termit.model.AbstractTerm;
-import cz.cvut.kbss.termit.model.FlatTerm;
 import cz.cvut.kbss.termit.model.Term;
 import cz.cvut.kbss.termit.model.Vocabulary;
 import cz.cvut.kbss.termit.model.util.HasIdentifier;
@@ -127,7 +125,7 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
         }
     }
 
-    private void postLoad(AbstractFullTerm r) {
+    private void postLoad(Term r) {
         r.setSubTerms(getSubTerms(r));
         r.setInverseRelated(loadInverseRelatedTerms(r));
         r.setInverseRelatedMatch(loadInverseRelatedMatchTerms(r));
@@ -139,7 +137,7 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
      *
      * @param term Term to load related terms for
      */
-    private Set<TermInfo> loadInverseRelatedTerms(AbstractFullTerm term) {
+    private Set<TermInfo> loadInverseRelatedTerms(Term term) {
         return loadInverseTermInfo(term, SKOS.RELATED,
                                    Utils.joinCollections(term.getRelated(), term.getRelatedMatch()));
     }
@@ -181,7 +179,7 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
      *
      * @param term Term to load related terms for
      */
-    private Set<TermInfo> loadInverseRelatedMatchTerms(AbstractFullTerm term) {
+    private Set<TermInfo> loadInverseRelatedMatchTerms(Term term) {
         return loadInverseTermInfo(term, SKOS.RELATED_MATCH, term.getRelatedMatch() != null ? term
                 .getRelatedMatch() : Collections.emptySet());
     }
@@ -191,7 +189,7 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
      *
      * @param term Term to load related terms for
      */
-    private Set<TermInfo> loadInverseExactMatchTerms(AbstractFullTerm term) {
+    private Set<TermInfo> loadInverseExactMatchTerms(Term term) {
         return loadInverseTermInfo(term, SKOS.EXACT_MATCH, term.getExactMatchTerms() != null ? term
                 .getExactMatchTerms() : Collections.emptySet());
     }
@@ -268,7 +266,6 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
      */
     private void evictPossiblyCachedReferences(Term term) {
         em.getEntityManagerFactory().getCache().evict(Term.class, term.getUri(), null);
-        em.getEntityManagerFactory().getCache().evict(FlatTerm.class, term.getUri(), null);
         em.getEntityManagerFactory().getCache().evict(TermDto.class, term.getUri(), null);
         em.getEntityManagerFactory().getCache().evict(TermInfo.class, term.getUri(), null);
         em.getEntityManagerFactory().getCache().evict(FlatTermDto.class, term.getUri(), null);
@@ -290,7 +287,6 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
           .setParameter("hasChild", URI.create(SKOS.NARROWER))
           .setParameter("t", term).getResultStream().forEach(st -> {
               em.getEntityManagerFactory().getCache().evict(Term.class, st, null);
-              em.getEntityManagerFactory().getCache().evict(FlatTerm.class, st, null);
               em.getEntityManagerFactory().getCache().evict(TermDto.class, st, null);
           });
     }
@@ -498,39 +494,6 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
     }
 
     /**
-     * Gets all terms from the specified vocabulary, regardless of their position in the term hierarchy.
-     * <p>
-     * Returns terms as a list of {@link FlatTerm} entities, i.e. only referencing direct parent terms.
-     *
-     * @param vocabulary Vocabulary whose terms should be returned
-     * @param pageSpec   Page specification
-     * @return Flat list of vocabulary terms
-     * @see #findAllFull(Vocabulary, Pageable)
-     */
-    public List<FlatTerm> findAllFullAndFlat(Vocabulary vocabulary, Pageable pageSpec) {
-        Objects.requireNonNull(vocabulary);
-        Objects.requireNonNull(pageSpec);
-        try {
-            // Load terms one by one. This works around the issue of terms being loaded in the persistence context
-            // as Term and TermInfo, which results in IndividualAlreadyManagedExceptions from JOPA
-            // The workaround relies on clearing the EntityManager after loading each term
-            // The price for this solution is that this method performs poorly for larger vocabularies (hundreds of terms)
-            final List<URI> termIris = findAllTermIris(vocabulary, pageSpec);
-            final Descriptor termDescriptor = descriptorFactory.termDescriptor(vocabulary);
-            return termIris.stream().map(ti -> {
-                final FlatTerm t = em.find(FlatTerm.class, ti, termDescriptor);
-                if (t != null) {
-                    postLoad(t);
-                    em.clear();
-                }
-                return t;
-            }).filter(Objects::nonNull).collect(Collectors.toList());
-        } catch (RuntimeException e) {
-            throw new PersistenceException(e);
-        }
-    }
-
-    /**
      * Gets all terms matching the specified search string from the specified vocabulary.
      * <p>
      * No differences are made between root terms and terms with parents. Note that this method returns terms with all
@@ -583,40 +546,6 @@ public class TermDao extends BaseAssetDao<Term> implements SnapshotProvider<Term
                  .setMaxResults(pageSpec.getPageSize())
                  .setFirstResult((int) pageSpec.getOffset())
                  .getResultList();
-    }
-
-    /**
-     * Gets all terms matching the specified search string from the specified vocabulary.
-     * <p>
-     * Returns terms as a list of {@link FlatTerm} entities, i.e. only referencing direct parent terms.
-     *
-     * @param searchString Search string to match term labels
-     * @param vocabulary   Vocabulary whose terms should be returned
-     * @param pageSpec     Page specification
-     * @return Flat list of vocabulary terms matching the search string
-     * @see #findAllFullAndFlat(Vocabulary, Pageable)
-     */
-    public List<FlatTerm> findAllFullAndFlat(String searchString, Vocabulary vocabulary, Pageable pageSpec) {
-        Objects.requireNonNull(vocabulary);
-        Objects.requireNonNull(pageSpec);
-        try {
-            // Load terms one by one. This works around the issue of terms being loaded in the persistence context
-            // as Term and TermInfo, which results in IndividualAlreadyManagedExceptions from JOPA
-            // The workaround relies on clearing the EntityManager after loading each term
-            // The price for this solution is that this method performs poorly for larger vocabularies (hundreds of terms)
-            final Descriptor termDescriptor = descriptorFactory.termDescriptor(vocabulary);
-            final List<URI> termIris = findAllTermIris(searchString, vocabulary, pageSpec);
-            return termIris.stream().map(ti -> {
-                final FlatTerm t = em.find(FlatTerm.class, ti, termDescriptor);
-                if (t != null) {
-                    postLoad(t);
-                    em.clear();
-                }
-                return t;
-            }).collect(Collectors.toList());
-        } catch (RuntimeException e) {
-            throw new PersistenceException(e);
-        }
     }
 
     private <T extends AbstractTerm> List<T> executeQueryAndLoadSubTerms(TypedQuery<T> query) {
