+++ src/main/java/cz/cvut/kbss/termit/service/business/TermService.java
@@ -149,9 +149,7 @@ public class TermService implements RudService<Term>, ChangeRecordProvider<Term>
         Objects.requireNonNull(selectionParams);
         if (selectionParams.full()) {
             // includeImported is not supported for finding full versions of terms
-            return selectionParams.flat() ?
-                   repositoryService.findAllFullAndFlat(vocabulary, selectionParams.pageSpec()) :
-                   repositoryService.findAllFull(vocabulary, selectionParams.pageSpec());
+            return repositoryService.findAllFull(vocabulary, selectionParams.pageSpec());
         } else {
             if (selectionParams.flat()) {
                 return selectionParams.includeImported() ?
@@ -182,9 +180,7 @@ public class TermService implements RudService<Term>, ChangeRecordProvider<Term>
         Objects.requireNonNull(selectionParams);
         if (selectionParams.full()) {
             // includeImported is not supported for finding full versions of terms
-            return selectionParams.flat() ?
-                   repositoryService.findAllFullAndFlat(searchString, vocabulary, selectionParams.pageSpec()) :
-                   repositoryService.findAllFull(searchString, vocabulary, selectionParams.pageSpec());
+            return repositoryService.findAllFull(searchString, vocabulary, selectionParams.pageSpec());
         } else {
             if (selectionParams.flat()) {
                 return selectionParams.includeImported() ?
